refactor: continued allocation tracker refactor into partitionable_devices.go
