+++ pkg/scheduling/dynamicresources/allocator.go
@@ -893,222 +893,6 @@ func computeTemplateTotals(slices []ResourceSlice) map[PoolKey]map[string]map[st
 	return totalsByPool
 }
 
-// poolCountersExhausted returns true if any counter in the pool has been fully consumed
-// by DFS-local tentative allocations, meaning no additional counter-consuming device from
-// this pool can succeed.
-func (a *allocator) poolCountersExhausted(pool *Pool) bool {
-	if len(pool.CounterSets) == 0 {
-		return false
-	}
-	remaining := a.allocationTracker.RemainingCounters[pool.Key]
-	if remaining == nil {
-		return false
-	}
-	allocating := a.allocatingCounters[pool.Key]
-	if allocating == nil {
-		return false
-	}
-	for counterSetName, counterSet := range allocating {
-		counterSetRemaining, ok := remaining[counterSetName]
-		if !ok {
-			continue
-		}
-		for counterName, allocCounter := range counterSet {
-			remCounter, ok := counterSetRemaining[counterName]
-			if !ok {
-				continue
-			}
-			if remCounter.Value.Value()-allocCounter.Value.Value() <= 0 {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// checkCounters verifies that shared counters have sufficient remaining budget for the device.
-// remainingCounterSets is the base budget (from AllocationTracker for in-cluster pools, or
-// templateRemainingCounters for template pools). The DFS-local allocatingCounters are subtracted
-// to account for tentative allocations in the current search.
-func (a *allocator) checkCounters(device cloudprovider.Device, poolKey PoolKey, remainingCounterSets map[string]map[string]resourcev1.Counter, template bool) bool {
-	if len(device.ConsumesCounters) == 0 {
-		return true
-	}
-	if remainingCounterSets == nil {
-		return false
-	}
-	allocatingCounterSets := lo.Ternary(template, a.templateAllocatingCounters[poolKey], a.allocatingCounters[poolKey])
-	for _, consumption := range device.ConsumesCounters {
-		counterSetRemaining, ok := remainingCounterSets[consumption.CounterSet]
-		if !ok {
-			return false
-		}
-		var allocatingCounters map[string]resourcev1.Counter
-		if allocatingCounterSets != nil {
-			allocatingCounters = allocatingCounterSets[consumption.CounterSet]
-		}
-		for counterName, counter := range consumption.Counters {
-			remainingCounter, ok := counterSetRemaining[counterName]
-			if !ok {
-				return false
-			}
-			allocatingVal := int64(0)
-			if allocatingCounters != nil {
-				if ac, ok := allocatingCounters[counterName]; ok {
-					allocatingVal = ac.Value.Value()
-				}
-			}
-			if remainingCounter.Value.Value()-allocatingVal < counter.Value.Value() {
-				return false
-			}
-		}
-	}
-	return true
-}
-
-// deductAllocatingCounters adds a device's counter consumption to the DFS-local allocating state.
-func (a *allocator) deductAllocatingCounters(device cloudprovider.Device, poolKey PoolKey, template bool) {
-	if len(device.ConsumesCounters) == 0 {
-		return
-	}
-	counterMap := lo.Ternary(template, a.templateAllocatingCounters, a.allocatingCounters)
-	allocatingCounterSets, ok := counterMap[poolKey]
-	if !ok {
-		allocatingCounterSets = make(map[string]map[string]resourcev1.Counter)
-		counterMap[poolKey] = allocatingCounterSets
-	}
-	for _, consumption := range device.ConsumesCounters {
-		allocatingCounters, ok := allocatingCounterSets[consumption.CounterSet]
-		if !ok {
-			allocatingCounters = make(map[string]resourcev1.Counter)
-			allocatingCounterSets[consumption.CounterSet] = allocatingCounters
-		}
-		for counterName, counter := range consumption.Counters {
-			allocatingCounter := allocatingCounters[counterName]
-			allocatingCounter.Value.Add(counter.Value)
-			allocatingCounters[counterName] = allocatingCounter
-		}
-	}
-}
-
-// restoreAllocatingCounters reverses a device's counter consumption from the DFS-local allocating state.
-func (a *allocator) restoreAllocatingCounters(device cloudprovider.Device, poolKey PoolKey, template bool) {
-	if len(device.ConsumesCounters) == 0 {
-		return
-	}
-	counterMap := lo.Ternary(template, a.templateAllocatingCounters, a.allocatingCounters)
-	allocatingCounterSets, ok := counterMap[poolKey]
-	if !ok {
-		return
-	}
-	for _, consumption := range device.ConsumesCounters {
-		allocatingCounters, ok := allocatingCounterSets[consumption.CounterSet]
-		if !ok {
-			continue
-		}
-		for counterName, counter := range consumption.Counters {
-			allocatingCounter, ok := allocatingCounters[counterName]
-			if !ok {
-				continue
-			}
-			allocatingCounter.Value.Sub(counter.Value)
-			allocatingCounters[counterName] = allocatingCounter
-		}
-	}
-}
-
-// countersFeasible checks whether the remaining counter budgets can possibly
-// satisfy the aggregate demand from all requests. This is a conservative
-// lower-bound check: if even the minimum total consumption exceeds available
-// budget, no DFS path can succeed. This is only done for AllMode requests
-// as their eligible devices (both in-cluster and template) are pre-computed.
-func (a *allocator) countersFeasible() bool {
-	for _, cd := range a.claimData {
-		for _, rd := range cd.Requests {
-			if rd.AllocationMode == resourcev1.DeviceAllocationModeAll {
-				if !a.allModeCountersFeasible(&rd) {
-					return false
-				}
-			}
-		}
-	}
-	return true
-}
-
-//nolint:gocyclo
-func (a *allocator) allModeCountersFeasible(rd *RequestData) bool {
-	// For All mode, we must allocate all predetermined devices.
-	// Decrement from shadow copies of remaining counters as we iterate.
-	// Map: poolKey -> counterSetName -> counterName -> Counter
-	inClusterShadow := make(map[PoolKey]map[string]map[string]resourcev1.Counter)
-	templateShadow := make(map[PoolKey]map[string]map[string]resourcev1.Counter)
-
-	devices := rd.AllDevices
-	if templateDevices, ok := rd.AllTemplateDevicesByIT[a.itID]; ok {
-		devices = append(devices, templateDevices...)
-	}
-
-	for _, d := range devices {
-		if len(d.ConsumesCounters) == 0 {
-			continue
-		}
-		poolKey := PoolKey{Driver: d.ID.Driver, Pool: d.ID.Pool}
-
-		var shadow map[PoolKey]map[string]map[string]resourcev1.Counter
-		if d.ID.Template {
-			shadow = templateShadow
-		} else {
-			shadow = inClusterShadow
-		}
-
-		if _, ok := shadow[poolKey]; !ok {
-			var remaining map[string]map[string]resourcev1.Counter
-			if d.ID.Template {
-				if a.templateRemainingCounters != nil {
-					remaining = a.templateRemainingCounters[poolKey]
-				}
-			} else {
-				remaining = a.allocationTracker.RemainingCounters[poolKey]
-			}
-			if remaining == nil {
-				return false
-			}
-			shadow[poolKey] = copyCounterSets(remaining)
-		}
-		poolShadow := shadow[poolKey]
-		for _, consumption := range d.ConsumesCounters {
-			counterSetsShadow, ok := poolShadow[consumption.CounterSet]
-			if !ok {
-				return false
-			}
-			for counterName, counter := range consumption.Counters {
-				availCounter, ok := counterSetsShadow[counterName]
-				if !ok {
-					return false
-				}
-				availCounter.Value.Sub(counter.Value)
-				if availCounter.Value.Cmp(resource.Quantity{}) < 0 {
-					return false
-				}
-				counterSetsShadow[counterName] = availCounter
-			}
-		}
-	}
-	return true
-}
-
-func copyCounterSets(src map[string]map[string]resourcev1.Counter) map[string]map[string]resourcev1.Counter {
-	cp := make(map[string]map[string]resourcev1.Counter, len(src))
-	for counterSetName, counters := range src {
-		cpCounters := make(map[string]resourcev1.Counter, len(counters))
-		for counterName, counter := range counters {
-			cpCounters[counterName] = resourcev1.Counter{Value: counter.Value.DeepCopy()}
-		}
-		cp[counterSetName] = cpCounters
-	}
-	return cp
-}
-
 // copyRequirements creates a shallow copy of a Requirements map.
 func copyRequirements(reqs scheduling.Requirements) scheduling.Requirements {
 	cp := scheduling.NewRequirements()
